Move signup and login forms into own directory
