+++ ui/client/src/App.js
@@ -6,8 +6,8 @@ import MenuContainer from "./components/MenuContainer";
 import LandingPage from './components/LandingPage.js';
 import UserForm from './components/UserForm.js';
 import ResumeForm from './components/resume';
-import SignUpForm from './components/SignUpForm.js';
-import LogInForm from './components/LogInForm.js';
+import SignUpForm from './components/Authentication/SignUpForm.js';
+import LogInForm from './components/Authentication/LogInForm.js';
 import ResumeInput from './components/ResumeInput.js';
 import PreviewPage from './components/PreviewPage.js'
 
@@ -23,9 +23,9 @@ export default function App() {
             <Route path='/users/new' element={<UserForm />} />
             <Route path='/users/:id/edit' element={<UserForm />} />
             <Route path='/resume/new' element={<ResumeForm />} />
-            <Route path='/signup' element={<SignUpForm onSignUp={() => {console.log("signed up")}} />} />
-            <Route path='/login' element={<LogInForm onLogIn={() => {console.log("Logged in")}} />} />
-            <Route path='/resume/datainput' element={<ResumeInput onSubmitResumeForm={() => {console.log("Submitted resume info")}} />} />
+            <Route path='/signup' element={<SignUpForm onSignUp={() => { console.log("signed up") }} />} />
+            <Route path='/login' element={<LogInForm onLogIn={() => { console.log("Logged in") }} />} />
+            <Route path='/resume/datainput' element={<ResumeInput onSubmitResumeForm={() => { console.log("Submitted resume info") }} />} />
             <Route path='/preview' element={<PreviewPage />} />
           </Routes>
         </Container>
